requirements.txt
Django>=4.0
psycopg2>=2.8
django-star-ratings